Adapt fastify
  - express 대신 fastify 로 바꿨습니다
  - 이외에 간단한 README 수정
  - 그리고 테스트 프로그램의 분리 등
