+++ .gitignore
@@ -1,3 +1,4 @@
+bin/
 lib/
 node_modules/
