+++ package.json
@@ -1,28 +1,31 @@
 {
   "name": "fake-iamport-server",
-  "version": "1.1.5",
+  "version": "1.2.0",
   "description": "Fake iamport server for testing",
   "main": "lib/index.js",
   "typings": "lib/index.d.ts",
   "scripts": {
     "----------------------------------------------": "",
     "build": "rimraf lib && tsc",
-    "build:api": "rimraf packages/api/lib && nestia sdk && npx copyfiles README.md packages/api && tsc -p tsconfig.api.json",
+    "build:api": "rimraf packages/api/lib && nestia sdk && npx copyfiles README.md packages/api && tsc -p packages/api/tsconfig.json",
     "build:sdk": "rimraf src/api/functional && nestia sdk",
     "build:swagger": "nestia swagger",
-    "dev": "npm run build -- --watch",
-    "eslint": "eslint ./src",
-    "eslint:fix": "eslint ./src --fix",
-    "prettier": "prettier src --write",
+    "build:test": "rimraf bin && tsc -p test/tsconfig.json",
+    "dev": "npm run build:test -- --watch",
+    "eslint": "eslint src && eslint --config .eslintrc.test.cjs test",
+    "eslint:fix": "eslint --fix src && eslint --fix --config .eslintrc.test.cjs test",
+    "prettier": "prettier src --write && prettier test --write",
     "------------------------------------------------": "",
     "package:api": "npm run build:api && cd packages/api && npm publish",
+    "package:latest": "npm run build && npm run build:test && npm run test && npm publish",
+    "package:next": "npm run package:latest -- --tag next",
     "prepare": "ts-patch install",
     "-------------------------------------------------": "",
     "start": "pm2 start lib/executable/server.js -i 1 --name fake-iamport-server --wait-ready --listen-timeout 120000 --kill-timeout 15000",
     "start:reload": "pm2 reload fake-iamport-server",
     "stop": "pm2 delete fake-iamport-server",
     "--------------------------------------------------": "",
-    "test": "node lib/test"
+    "test": "node bin/test"
   },
   "repository": {
     "type": "git",
@@ -35,34 +38,33 @@
   },
   "homepage": "https://github.com/samchon/fake-iamport-server",
   "devDependencies": {
-    "@nestia/sdk": "^1.0.17",
+    "@nestia/sdk": "^1.2.5",
     "@trivago/prettier-plugin-sort-imports": "^4.0.0",
     "@types/atob": "^2.1.2",
     "@types/btoa": "^1.2.3",
     "@types/cli": "^0.11.19",
-    "@types/express": "^4.17.12",
     "@types/node": "^15.6.1",
     "@types/uuid": "^8.3.0",
     "@typescript-eslint/eslint-plugin": "^5.26.0",
     "@typescript-eslint/parser": "^5.26.0",
     "cli": "^1.0.1",
     "copyfiles": "^2.4.1",
-    "nestia": "^4.0.10",
+    "nestia": "^4.2.0",
     "pm2": "^4.5.6",
     "prettier": "^2.6.2",
     "rimraf": "^3.0.2",
     "sloc": "^0.2.1",
-    "ts-node": "^10.8.0",
+    "ts-node": "^10.9.1",
     "ts-patch": "^2.1.0",
     "typescript": "^4.9.5"
   },
   "dependencies": {
-    "@nestia/core": "^1.0.17",
-    "@nestia/e2e": "^0.1.5",
+    "@nestia/core": "^1.2.2",
+    "@nestia/e2e": "^0.2.2",
     "source-map-support": "^0.5.19",
     "tstl": "^2.5.13",
     "typescript-transform-paths": "^3.4.6",
-    "typia": "^3.6.9",
+    "typia": "^3.8.6",
     "uuid": "^8.3.2"
   },
   "keywords": [
@@ -78,8 +80,6 @@
     "README.md",
     "LICENSE",
     "lib",
-    "src",
-    "!lib/test",
-    "!src/test"
+    "src"
   ]
 }
