+++ packages/api/tsconfig.json
@@ -14,10 +14,10 @@
       // "checkJs": true,                       /* Report errors in .js files. */
       // "jsx": "preserve",                     /* Specify JSX code generation: 'preserve', 'react-native', or 'react'. */
       "declaration": true,                   /* Generates corresponding '.d.ts' file. */
-      "declarationMap": true,                /* Generates a sourcemap for each corresponding '.d.ts' file. */
-      // "sourceMap": true,                     /* Generates corresponding '.map' file. */
+      // "declarationMap": true,                /* Generates a sourcemap for each corresponding '.d.ts' file. */
+      "sourceMap": true,                     /* Generates corresponding '.map' file. */
       // "outFile": "./",                       /* Concatenate and emit output to single file. */
-      "outDir": "./packages/api/lib",                        /* Redirect output structure to the directory. */
+      "outDir": "./lib",                        /* Redirect output structure to the directory. */
       // "rootDir": "./",                       /* Specify the root directory of input files. Use to control the output directory structure with --outDir. */
       // "composite": true,                     /* Enable project compilation */
       // "tsBuildInfoFile": "./",               /* Specify file to store incremental compilation information */
@@ -70,5 +70,5 @@
       "skipLibCheck": true,                     /* Skip type checking of declaration files. */
       "forceConsistentCasingInFileNames": true  /* Disallow inconsistently-cased references to the same file. */
     },
-    "include": ["src/api"]
+    "include": ["../../src/api"]
   }
