+++ src/api/functional/certifications/otp/index.ts
@@ -115,6 +115,6 @@ export namespace confirm
 
     export function path(imp_uid: string): string
     {
-        return `/certifications/otp/confirm/${encodeURIComponent(imp_uid)}`;
+        return `/certifications/otp/confirm/${encodeURIComponent(imp_uid ?? "null")}`;
     }
 }
