+++ src/api/functional/payments/index.ts
@@ -52,7 +52,7 @@ export namespace at
 
     export function path(imp_uid: string): string
     {
-        return `/payments/${encodeURIComponent(imp_uid)}`;
+        return `/payments/${encodeURIComponent(imp_uid ?? "null")}`;
     }
 }
